+++ argueview/ArgueView.py
@@ -1,6 +1,6 @@
-from typing import List, Union, Dict, Tuple
+from typing import List, Union
 from argueview.typings import Source, Explanation, Case, Data, ExplanationPartial, FeatureExplanation, Grounds, \
-    GroundsVarByClass, FeatureImportance, Feature
+    GroundsVarByClass, FeatureImportance, Feature, LatentContinuousTarget
 
 
 class ArgueView:
@@ -14,6 +14,7 @@ class ArgueView:
     _backing: str
     _sources: List[Source]
     _classes: List[str]
+    _lct: LatentContinuousTarget
     _grounds: Grounds
     _grounds_vars: GroundsVarByClass
 
@@ -37,8 +38,30 @@ class ArgueView:
         :param  classes: Array of textual representations for your decision-classes.
         :return:
         """
-
+        if self._lct and len(self._lct.mapping) != len(classes):
+            raise Exception('The number of classes should match the length of the defined latent continuous target mapping.')
         self._classes = classes
+
+    def latent_continuous_target(self, label: str, anti_label: str, mapping: List[float]) -> None:
+        """Define a latent continuous target present within your decision-classes.
+
+        To illustrate the usage of the latent continuous target assume a decision task in the loan application domain
+        where predictions determine whether someone is eligible for a loan. The decision-classes are: 'applicable for loan'
+        and 'not applicable for loan'. In this scenario the latent continuous target is 'applicability' and it's anti-label
+        is 'inapplicability'. The mapping is [1, -1] which indicates the value of the latent target for each decision-class.
+
+        :param label: Label for the latent continuous target.
+        :param anti_label: Anti-label for the latent continuous target.
+        :param mapping: List of values representing the value of the latent target for each decision-class.
+        :return:
+        """
+        if self._classes and len(self._classes) != len(mapping):
+            raise Exception('Length of mapping should match the number of classes.')
+        self._lct = LatentContinuousTarget({
+            label: label,
+            anti_label: anti_label,
+            mapping: mapping
+        })
 
     def backing(self, backing: str) -> None:
         """Defines textual backing to support your decisions.
@@ -83,11 +106,11 @@ class ArgueView:
         ground = self._grounds[feature][decision]
         feature_value = int(self._case.sources[source].features[feature].value)
         options = self._grounds_vars[feature]
-        if (feature_value >= len(options)):
+        if feature_value >= len(options):
             return ground
         return ground.replace("<>", options[feature_value])
 
-    def _generateExplanationPartial(self) -> None:
+    def _generate_explanation_partial(self) -> None:
         positive = []
         negative = []
         decision = self._case.decision_class()
@@ -109,6 +132,19 @@ class ArgueView:
         })
 
     def generate(self, case: Case, feature_importance: Union[FeatureImportance, List[FeatureImportance]], unexplained: float = 0) -> Explanation:
+        """Generate an explanation for a case based on the given context.
+
+        Generates an explanation for a specific case. This method uses the information provided by other methods and
+        assembles this information and the current case into an explanation object. This object can be used to
+        visualize the explanation or to export it for usage in other tools.
+
+        :param case: Case object containing the class probabilities and feature values.
+        :param feature_importance: Feature importance for the case as computed by an explainer. A list of FeatureImportance
+        objects is required for multi-source scenarios.
+        :param unexplained: Unexplained contribution for the decision-class.
+        :return:
+        """
+
         pre = 'cannot generate explanation: '
         if not self._classes:
             raise Exception(pre+'no classes defined. Use ArgueView.classes to set your classes.')
@@ -126,7 +162,7 @@ class ArgueView:
         self._unexplained = unexplained
 
         self._case = case
-        self._generateExplanationPartial()
+        self._generate_explanation_partial()
         _gen = Explanation()
 
         if self._backing:
@@ -135,6 +171,7 @@ class ArgueView:
         _gen.data = Data()
         _gen.data.classes = self._classes
         _gen.data.sources = self._sources
+        _gen.data.latent_continuous_target = self._lct
         _gen.case = self._case
         _gen.explanation = self._expPartial
         return _gen