+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 
 ### v0.1.2
  - fix explanation exporting in json
+ 
+### v0.1.3
+ - add Toulmin visualizer
+ 
+### v0.1.4
+ - add latent continuous target
+ - complement documentation